+++ bindings/mnt/pyfiction/test/layouts/test_bounding_box.py
@@ -30,18 +30,6 @@ from mnt.pyfiction import (
             lambda: shifted_cartesian_gate_layout((2, 2, 0), "2DDWave", "Layout"), id="shifted_cartesian_gate_layout"
         ),
         pytest.param(lambda: hexagonal_gate_layout((2, 2, 0), "2DDWave", "Layout"), id="hexagonal_gate_layout"),
-        pytest.param(
-            lambda: cartesian_gate_layout((2, 2, 0), "2DDWave", "Layout"),
-            id="cartesian_obstruction_layout",
-        ),
-        pytest.param(
-            lambda: shifted_cartesian_gate_layout((2, 2, 0), "2DDWave", "Layout"),
-            id="shifted_cartesian_obstruction_layout",
-        ),
-        pytest.param(
-            lambda: hexagonal_gate_layout((2, 2, 0), "2DDWave", "Layout"),
-            id="hexagonal_obstruction_layout",
-        ),
     ],
 )
 def test_bounding_box_around_an_empty_gate_level_layout(make_layout):