+++ bindings/mnt/pyfiction/test/physical_design/path_finding/test_distance.py
@@ -27,15 +27,12 @@ from mnt.pyfiction import (
 
 ALL_LAYOUTS = [
     pytest.param(lambda: cartesian_layout((4, 4)), id="cartesian_layout"),
-    pytest.param(lambda: cartesian_gate_layout((4, 4), "2DDWave"), id="cartesian_gate_layout"),
     pytest.param(lambda: cartesian_gate_layout((4, 4), "2DDWave", "Layout"), id="cartesian_gate_layout"),
     pytest.param(lambda: shifted_cartesian_layout((4, 4)), id="shifted_cartesian_layout"),
-    pytest.param(lambda: shifted_cartesian_gate_layout((4, 4), "2DDWave"), id="shifted_cartesian_gate_layout"),
     pytest.param(
         lambda: shifted_cartesian_gate_layout((4, 4), "2DDWave", "Layout"), id="shifted_cartesian_gate_layout"
     ),
     pytest.param(lambda: hexagonal_layout((4, 4)), id="hexagonal_layout"),
-    pytest.param(lambda: hexagonal_gate_layout((4, 4), "2DDWave"), id="hexagonal_gate_layout"),
     pytest.param(lambda: hexagonal_gate_layout((4, 4), "2DDWave", "Layout"), id="hexagonal_gate_layout"),
 ]